+++ python/src/lib.rs
@@ -53,6 +53,12 @@ fn spaceweather(_py: Python, m: &Bound<'_, PyModule>) -> PyResult<()> {
         .unwrap();
     m.add_function(wrap_pyfunction!(pyspaceweather::predicted_f107, m)?)
         .unwrap();
+    m.add_function(wrap_pyfunction!(pyspaceweather::coverage, m)?)?;
+    m.add_function(wrap_pyfunction!(pyspaceweather::status, m)?)?;
+    m.add_function(wrap_pyfunction!(
+        pyspaceweather::disable_space_weather_time_warning,
+        m
+    )?)?;
     m.add_function(wrap_pyfunction!(pyspaceweather::update, m)?)
         .unwrap();
     Ok(())